+++ examples/waku/src/pages/docs/_layout.tsx
@@ -1,20 +1,11 @@
 import type { ReactNode } from 'react';
 import { DocsLayout } from 'fumadocs-ui/layouts/docs';
 import { source } from '@/lib/source';
+import { baseOptions } from '@/lib/layout.shared';
 
-export default function Layout({
-  children,
-}: {
-  children: ReactNode;
-  path: string;
-}) {
+export default function Layout({ children }: { children: ReactNode }) {
   return (
-    <DocsLayout
-      tree={source.pageTree}
-      nav={{
-        title: 'Waku',
-      }}
-    >
+    <DocsLayout {...baseOptions()} tree={source.pageTree}>
       {children}
     </DocsLayout>
   );