+++ packages/create-app/template/waku/src/pages/(home)/_layout.tsx
@@ -1,0 +1,13 @@
+import type { ReactNode } from 'react';
+import { baseOptions } from '@/lib/layout.shared';
+import { HomeLayout } from 'fumadocs-ui/layouts/home';
+
+export default function Layout({ children }: { children: ReactNode }) {
+  return <HomeLayout {...baseOptions()}>{children}</HomeLayout>;
+}
+
+export const getConfig = async () => {
+  return {
+    render: 'static',
+  };
+};