+++ packages/typescript/src/markdown.ts
@@ -1,4 +1,4 @@
-import type { Nodes } from 'hast';
+import type { ElementContent, Nodes } from 'hast';
 import { remark } from 'remark';
 import {
   rehypeCode,
@@ -6,31 +6,48 @@ import {
   remarkGfm,
 } from 'fumadocs-core/mdx-plugins';
 import remarkRehype from 'remark-rehype';
+import { highlightHast } from 'fumadocs-core/highlight';
+
+const shikiOptions = {
+  lazy: true,
+
+  themes: {
+    light: 'github-light',
+    dark: 'github-dark',
+  },
+} satisfies RehypeCodeOptions;
 
 const processor = remark()
   .use(remarkGfm)
   .use(remarkRehype)
-  .use(rehypeCode, {
-    lazy: true,
+  .use(rehypeCode, shikiOptions);
 
-    themes: {
-      light: 'github-light',
-      dark: 'github-dark',
-    },
-  } satisfies RehypeCodeOptions)
-  // @ts-expect-error -- safe
-  .use(() => {
-    return (tree, file: { data: Record<string, unknown> }) => {
-      file.data.tree = tree;
-
-      return '';
-    };
+export async function renderTypeToHast(type: string): Promise<Nodes> {
+  const nodes = await highlightHast(type, {
+    ...shikiOptions,
+    lang: 'ts',
+    structure: 'inline',
   });
+
+  return {
+    type: 'element',
+    tagName: 'span',
+    properties: {
+      class: 'shiki',
+    },
+    children: [
+      {
+        type: 'element',
+        tagName: 'code',
+        properties: {},
+        children: nodes.children as ElementContent[],
+      },
+    ],
+  };
+}
 
 export async function renderMarkdownToHast(md: string): Promise<Nodes> {
   md = md.replace(/{@link (?<link>[^}]*)}/g, '$1'); // replace jsdoc links
 
-  const out = await processor.process(md);
-
-  return out.data.tree as Nodes;
+  return processor.run(processor.parse(md));
 }