+++ packages/typescript/CHANGELOG.md
@@ -1,5 +1,12 @@
 # fumadocs-typescript
 
+## 4.0.7
+
+### Patch Changes
+
+- 45c7531: Type Table: Support displaying parameters & return types
+- 4082acc: Redesign Type Table
+
 ## 4.0.6
 
 ### Patch Changes
@@ -158,7 +165,6 @@
 - f75287d: **Introduce `fumadocs-docgen` package.**
 
   Offer a better authoring experience for advanced use cases.
-
   - Move `remark-dynamic-content` and `remark-install` plugins to the new package `fumadocs-docgen`.
   - Support Typescript generator by default